Move unit tests requiring access to data on disk to integration
