+++ pm_icecon/tests/unit/test_bt_params.py
@@ -14,7 +14,6 @@ print(_get_config_hash(new_params))
 
 Then setup a new test that asserts that hash doesn't change!
 """
-import datetime as dt
 import hashlib
 import json
 from pathlib import Path
@@ -22,8 +21,6 @@ from pathlib import Path
 import numpy as np
 
 import pm_icecon.bt.params.ausi_amsr2 as amsr2_params
-from pm_icecon.bt.fields import get_bootstrap_fields
-from pm_icecon.bt.params.util import convert_to_pmicecon_bt_params
 
 
 # This class stolen from qgreenland's config export code...
@@ -86,29 +83,3 @@ def test_cdr_amsr2_params():
         _get_config_hash(amsr2_params.CDR_AMSR2_SOUTH_PARAMS)
         == 'f9bd22a2d2c48a874d89c51cb2480436'
     )
-
-
-def test_ausi12_amsr2_bt_params_north():
-    date = dt.date(2022, 1, 1)
-    fields = get_bootstrap_fields(date=date, satellite='amsr2', gridid='e2ns25')
-    params = amsr2_params.get_ausi_bootstrap_params(
-        date=date, satellite='amsr2', gridid='e2ns25'
-    )
-    bt_params = convert_to_pmicecon_bt_params(
-        hemisphere='north', params=params, fields=fields
-    )
-
-    assert _get_config_hash(bt_params) == '8e61f93a2f762e962323f159342d282c'
-
-
-def test_ausi12_amsr2_bt_params_south():
-    date = dt.date(2022, 1, 1)
-    fields = get_bootstrap_fields(date=date, satellite='amsr2', gridid='e2ss25')
-    params = amsr2_params.get_ausi_bootstrap_params(
-        date=date, satellite='amsr2', gridid='e2ss25'
-    )
-    bt_params = convert_to_pmicecon_bt_params(
-        hemisphere='south', params=params, fields=fields
-    )
-
-    assert _get_config_hash(bt_params) == 'a5e1d2959f31ad165a075cd9003fbdff'
